Multiplatform setup script

+++ setup.sh
@@ -18,16 +18,32 @@ case "$1" in
     ;;
 esac
 
+#detect platform
+platform='unknown'
+unamestr=`uname`
+if [[ "$unamestr" == 'Linux' ]]; then
+   platform='linux'
+elif [[ "$unamestr" == 'Darwin' ]]; then
+   platform='osx'
+else
+	echo 'Unknown platform, aborting.'
+	exit
+fi
+
 ## should check if you do not have wget
 #fix folder layout
 
 # ******************************* #
 # * JLINK Deployer and Debugger * #
 # ******************************* #
-mkdir $DEPLOY_RESOURCES/jlink
-wget --post-data="agree=1&confirm=yes" "https://www.segger.com/jlink-software.html?step=1&file=JLinkMacOSX_502f" -O $DEPLOY_RESOURCES/jlink/JLink_MacOSX_V502f.pkg
-# should check checksum
-open $DEPLOY_RESOURCES/jlink/JLink_MacOSX_V502f.pkg
+if [[ $platform == 'osx' ]]; then
+	mkdir $DEPLOY_RESOURCES/jlink
+	wget --post-data="agree=1&confirm=yes" "https://www.segger.com/jlink-software.html?step=1&file=JLinkMacOSX_502f" -O $DEPLOY_RESOURCES/jlink/JLink_MacOSX_V502f.pkg
+	# should check checksum
+	open $DEPLOY_RESOURCES/jlink/JLink_MacOSX_V502f.pkg
+elif [[ $platform == 'linux' ]]; then	
+	echo 'For linux, manually install segger JLink tools version 9 from www.segger.com'
+fi
 
 # ********************* #
 # * GCC ARM Toolchain * #
@@ -35,10 +51,17 @@ open $DEPLOY_RESOURCES/jlink/JLink_MacOSX_V502f.pkg
 mkdir $GCC_ARM_TOOLCHAIN
 # should check checksum
 # should get the latest gcc-arm-none-eabi
-wget -O $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4-20151219-mac.tar.bz2 https://launchpad.net/gcc-arm-embedded/5.0/5-2015-q4-major/+download/gcc-arm-none-eabi-5_2-2015q4-20151219-mac.tar.bz2
-tar -C $GCC_ARM_TOOLCHAIN -xjf $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4-20151219-mac.tar.bz2
-mv $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4/* $GCC_ARM_TOOLCHAIN
-rm -r $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4-20151219-mac.tar.bz2 $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4/
+if [[ $platform == 'osx' ]]; then
+	wget -O $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4-20151219-mac.tar.bz2 https://launchpad.net/gcc-arm-embedded/5.0/5-2015-q4-major/+download/gcc-arm-none-eabi-5_2-2015q4-20151219-mac.tar.bz2
+	tar -C $GCC_ARM_TOOLCHAIN -xjf $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4-20151219-mac.tar.bz2
+	mv $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4/* $GCC_ARM_TOOLCHAIN
+	rm -r $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4-20151219-mac.tar.bz2 $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4/
+elif [[ $platform == 'linux' ]]; then
+	wget -O $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4-20151219-mac.tar.bz2 https://launchpad.net/gcc-arm-embedded/5.0/5-2015-q4-major/+download/gcc-arm-none-eabi-5_2-2015q4-20151219-linux.tar.bz2
+	tar -C $GCC_ARM_TOOLCHAIN -xjf $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4-20151219-linux.tar.bz2
+	mv $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4/* $GCC_ARM_TOOLCHAIN
+	rm -r $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4-20151219-linux.tar.bz2 $GCC_ARM_TOOLCHAIN/gcc-arm-none-eabi-5_2-2015q4/	
+fi	
 
 # ********************* #
 # * Nordic Softdevice * #
@@ -69,20 +92,22 @@ cp $DEPLOY_RESOURCES/nrf_drv_config.h.fix $BAD_CONFIG_FILE
 # ************************* #
 # * cmocka test framework * #
 # ************************* #
-mkdir $CMOCKA
-wget -O $CMOCKA/cmocka-1.0.1.tar.xz https://cmocka.org/files/1.0/cmocka-1.0.1.tar.xz
-tar -C $CMOCKA -xvf $CMOCKA/cmocka-1.0.1.tar.xz
-mv $CMOCKA/cmocka-1.0.1/* $CMOCKA
-mv $CMOCKA/cmocka-1.0.1/.[!.]* $CMOCKA
-rmdir $CMOCKA/cmocka-1.0.1
-rm $CMOCKA/cmocka-1.0.1.tar.xz
-mkdir $CMOCKA/build
-cd $CMOCKA/build
-cmake -DCMAKE_INSTALL_PREFIX=/usr -DCMAKE_BUILD_TYPE=Debug ..
-make
-echo 'Installing cmocka test framework libraries under /usr/lib'
-sudo make install
-cd -
+if [[ $platform == 'linux' ]]; then
+	mkdir $CMOCKA
+	wget -O $CMOCKA/cmocka-1.0.1.tar.xz https://cmocka.org/files/1.0/cmocka-1.0.1.tar.xz
+	tar -C $CMOCKA -xvf $CMOCKA/cmocka-1.0.1.tar.xz
+	mv $CMOCKA/cmocka-1.0.1/* $CMOCKA
+	mv $CMOCKA/cmocka-1.0.1/.[!.]* $CMOCKA
+	rmdir $CMOCKA/cmocka-1.0.1
+	rm $CMOCKA/cmocka-1.0.1.tar.xz
+	mkdir $CMOCKA/build
+	cd $CMOCKA/build
+	cmake -DCMAKE_INSTALL_PREFIX=/usr -DCMAKE_BUILD_TYPE=Debug ..
+	make
+	echo 'Installing cmocka test framework libraries under /usr/lib'
+	sudo make install
+	cd -
+fi
 
 ### BELOW DOES NOT YET WORK ###
